fix(lib): move mfa logic from sessionStorage to localStorage
* There were some flags stored in sessionStorage which allow to handle mfa logic correctly across
windows. The idea was, that the sessionStorage is copied between windows when the new window is opended
via window.open(). It works in Chrome and Firefox, but e.g. in Vivaldi the behavior is different and
the sessionStorage is not copied. According to this the better way is to store these flags in localStorage.