fix(profile): navigation working properly in mailing lists settings
Fixed a problem of getting stuck in the mailing lists settings.
When navigating away from the component it clears the query params
and replaces the old url. The query params need to be removed also from
the new component. This is done in the onDestroyed method of the mailing lists.

(cherry picked from commit 1df1e65d9164aa38da0f041636643b3a1e01bd8e)