fix(profile): fix the setting of MFA and move the whole logic to step-up
* Previously, when the user was authenticated by just one factor and wanted to change MFA settings,
he/she was forced to completely log out from application and again log in with both factors.
* Now this logic is handled by shared logic for step-up. It is partially customized for this use case
(it is not dependent on instance configuration etc.), but much logic/code is reused for this purpose.
* Some code became redundant and useless, so the component and service were refactored a bit.
* Some helper functions in the mfa-api service were rewrote to rxjs.

DEPLOYMENT NOTE: New property `auth_validity` for user-profile under the mfa has been created.
This property defines for how long MFA will not be required since the last valid MFA (only for MFA settings).