fix: reworked links to other apps
- links to other apps in the header are now set in config
- possible to have more than one link or none
- also fixed the main content size in the publications app

BREAKING_CHANGE: Config properties `profile_label(_cs)` and `admin_gui_label`
are removed and replaced by a new property `other_apps`. On first level it
maps a language to a new map. This new map then maps the keys
(one of `admin`, `profile`, `pwdReset`, `consolidator`, `linker`) to a
label which will be displayed in the other apps submenu.